backend/requirements.txt
aiosqlite==0.19.0
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.7
httpx==0.25.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.24.0
